Refactor: 이미지 로직 코드 분리
